Modularize crate
